moved NewChartWidget to base, updated paths & removed component here
