+++ src/js/components/pages/greendash/TimeOfDayCard.jsx
@@ -4,7 +4,7 @@ import _ from 'lodash';
 
 import DataStore from '../../../base/plumbing/DataStore';
 import Misc from '../../../base/components/Misc';
-import NewChartWidget from '../../NewChartWidget';
+import NewChartWidget from '../../../base/components/NewChartWidget';
 import { dataColours, GreenCard, GreenCardAbout, NOEMISSIONS, TONNES_THRESHOLD } from './dashutils';
 import { getBreakdownBy, getCarbon } from './carboncalc';
